1)added plant type to list of plants, 2)removed delete image button and made an option, 3) refactor code

+++ app/src/main/java/com/example/teamg_plantproject/ImageDisplayActivity.java
@@ -1,66 +1,71 @@
 package com.example.teamg_plantproject;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ImageDisplayActivity extends AppCompatActivity {
     protected String path;
     protected TextView date;
-    protected Button deleteImgButton;
     protected String sensorID;
     protected DatabaseHelper db;
     int pictureNumber;
     Image image;
     String image_date;
+    private int plantId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_image_display);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
-        deleteImgButton = findViewById(R.id.delete_img_button);
-        ImageView imageView = (ImageView)findViewById(R.id.imageView);
+        ImageView imageView = findViewById(R.id.imageView);
         date = findViewById(R.id.dateStamp);
 
-        image= getIntent().getParcelableExtra("Image");
+        image = getIntent().getParcelableExtra("Image");
+        plantId = getIntent().getIntExtra("PlantID", 0);
         imageView.setImageBitmap(image.getImage());
 
         date.setText(image.getImageDate());
-        Log.d("TAG", "onCreate: IMAGE NUMBER: "+ image.getImageNumber());
+        Log.d("TAG", "onCreate: IMAGE NUMBER: " + image.getImageNumber());
         pictureNumber = image.getImageNumber();
-
-        deleteImgButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                //Deleting records from database
-                db = new DatabaseHelper(getApplicationContext());
-                db.deletePlantPicture(pictureNumber);
-                //after deleting data go to main page
-                Intent intent = new Intent(ImageDisplayActivity.this, ImageArchive.class);
-                startActivity(intent);
-                finish();
-            }
-        });
     }
+
     @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater menuInflater = getMenuInflater();
+        menuInflater.inflate(R.menu.individual_image_options, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        if (item.getItemId() == R.id.option_delete_image) {
+            db = new DatabaseHelper(getApplicationContext());
+            db.deletePlantPicture(pictureNumber);
+            //after deleting data go to main page
+            Intent intent = new Intent(this, ImageArchive.class);
+            System.out.println(plantId + "++++++++++++++++++++++++++++" + "delete");
+//            intent.putExtra("PlantId", plantId);
+            startActivity(intent);
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
     public boolean onSupportNavigateUp() {
-        Intent intent = new Intent(this.getApplicationContext(), ImageArchive.class);
-        startActivity(intent);
+        Intent intent = new Intent(this, ImageArchive.class);
+        intent.putExtra("PlantID", plantId);
+        System.out.println(plantId + "++++++++++++++++++++++++++++" + "back");
+        this.startActivity(intent);
         return true;
     }
 }
